Finish recipe function
Finished the recipe batch. I believe the time complexity if O(n). I think I did
pretty well on having this one work properly first go. I even have test cases
to immediatley end the problem if the lengths are not the same, or a recipe
and ingrediant key do not match.